Extract InstructionBlockBaseNode into lib/editor/
Extract the pure schema and markdown parsing logic from
InstructionBlockExtension into a separate base node in lib/editor/.
InstructionBlockExtension now extends the base node with UI-only behavior.
